add class KerberosApacheAuth and example code
